Introduces ValidatorChain functionality
Adds a `ValidatorChain`. It only allows attaching validator instances,
and only in the order in which they should execute.
`$breakChainOnFailure` is preserved.

In order to return results, this introduces a `ResultAggregate` as well,
which accepts the value to the constructor; the chain then pushes
results onto it. Its `isValid()` and `getMessages()` functionality use
`array_reduce` operations in order to provide the results of validation.