Refactor Barcode validators
Extracts checksum functionality to a new interface,
`ChecksummableInterface`, which defines the various methods for
determining if a checksum is in use, as well as how to invoke it.

Most checksummable barcode validators now accept a boolean flag,
`$useChecksum` to its constructor, defaulting it to whatever the default
behavior should be (generally `true`, indicating checksum validation
should occur). Some override `useChecksum()` to determine if the value
requires a checksum validation, usually based on length or an initial or
final character in the barcode.

**ALL** barcode validators now have only constructor injection for
defining behavior. This ensures they are always in a consistent state.

Finally, all barcode validators now have strict_types enabled, which
caught a number of errors in how the various checksum operations were
being calculated and compared.